Rename pkg/store/metrics -> pkg/store/telemetry
